SEQNG-468 Split the handlers package into several files
